server/app.py
import os
from dotenv import load_dotenv

load_dotenv()

from flask import Flask
from flask_cors import CORS
from routes.health_routes import health_bp
from routes.chat_routes import chat_bp

app = Flask(__name__)
CORS(app)

app.register_blueprint(health_bp)
app.register_blueprint(chat_bp)

if __name__ == '__main__':
    port = int(os.getenv('PORT', 5000))
    app.run(host='0.0.0.0', port=port, debug=True) 