server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
langchain==0.1.12
langchain-community>=0.0.28
langchain-google-genai==0.0.11
google-generativeai>=0.4.1,<0.5.0
supabase==1.2.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9 